requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
openai-whisper>=20221001
torch>=1.9.0
torchaudio>=0.9.0
ffmpeg-python>=0.2.1
numpy>=1.21.0
